backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
xgboost
shap
python-whois
joblib
python-dotenv
 requests
